requirements.txt
beautifulsoup4~=4.12.2
pandas~=2.0.3
selenium==4.11.2
requests~=2.31.0
PyYAML==6.0.1
urllib3==2.0.5